refactor: migrate TerminalRenderer logic to TerminalPlugin hooks
Move ANSI rendering, syntax highlighting, markdown inline conversion,
DSL rendering, and all ContentBlock rendering from TerminalRenderer
into TerminalPlugin. TerminalPlugin now overrides IMPlugin hook methods
(render_code_block, render_markdown, render_hr) and owns the ansi
flag directly. Deleted src/renderer/terminal.rs and its tests.

Source: user
Type: refactor
