+++ src/im_adapter/renderer.rs
@@ -1,6 +1,5 @@
 // Renderer trait has been removed as part of the IMPlugin refactoring.
 // RenderedOutput is now defined in [`crate::im_adapter::plugin`].
 //
-// This file is intentionally left minimal. Platform-specific renderers
-// (FeishuRenderer, TerminalRenderer) still live in `src/renderer/` and will
-// be migrated in subsequent steps.
+// Platform-specific rendering logic (Feishu, Terminal) now lives in the
+// corresponding IMPlugin implementations under `src/im/`.
